Chat Fragment fixed to scroll to latest chat! another crash while adding user in a created trip fixed.

+++ Travelogue/app/src/main/java/sai/developement/travelogue/fragments/TripDetailsFragment.java
@@ -165,7 +165,7 @@ public class TripDetailsFragment extends Fragment {
                                         user.setEmail((String) ((HashMap) pair.getValue()).get("email"));
                                         user.setId((String) ((HashMap) pair.getValue()).get("id"));
                                         user.setName((String) ((HashMap) pair.getValue()).get("name"));
-                                        user.setAvatarId((Integer)  ((HashMap) pair.getValue()).get("avatarId"));
+                                        user.setAvatarId(((Long)((HashMap) pair.getValue()).get("avatarId")).intValue());
                                         break;
                                     }
